+++ src/frontend/apps/impress/src/components/Title.tsx
@@ -1,21 +1,21 @@
 import { useTranslation } from 'react-i18next';
 
-import { Text } from '@/components/';
+import { Text, TextType } from '@/components/';
 
-type TitleSemanticsProps = {
+type TitleSemanticsProps = TextType & {
   headingLevel?: 'h1' | 'h2' | 'h3';
   className?: string;
 };
 
 export const Title = ({
   headingLevel = 'h2',
-  className,
+  ...props
 }: TitleSemanticsProps) => {
   const { t } = useTranslation();
 
   return (
     <Text
-      className={`--docs--title${className ? ` ${className}` : ''}`}
+      className={`--docs--title${props.className ? ` ${props.className}` : ''}`}
       $direction="row"
       $align="center"
       $margin="none"
@@ -23,6 +23,7 @@ export const Title = ({
       $zIndex={1}
       $size="1.375rem"
       $color="var(--c--contextuals--content--logo1)"
+      {...props}
     >
       {t('Docs')}
     </Text>