+++ src/frontend/apps/impress/src/components/separators/SeparatedSection.tsx
@@ -1,25 +1,23 @@
 import { PropsWithChildren } from 'react';
 import { css } from 'styled-components';
 
-import { useCunninghamTheme } from '@/cunningham';
+import { Box, BoxType } from '../Box';
 
-import { Box } from '../Box';
-
-type SeparatedSectionProps = {
+type SeparatedSectionProps = BoxType & {
   showSeparator?: 'top' | 'bottom' | boolean;
 };
 
 export const SeparatedSection = ({
   showSeparator = true,
   children,
+  ...boxProps
 }: PropsWithChildren<SeparatedSectionProps>) => {
-  const { spacingsTokens } = useCunninghamTheme();
-
   return (
     <Box
+      $width="100%"
+      $padding={{ vertical: 'sm', horizontal: '0' }}
+      {...boxProps}
       $css={css`
-        width: 100%;
-        padding: ${spacingsTokens['sm']} 0;
         ${showSeparator === 'top' || showSeparator === true
           ? css`
               border-top: 1px solid