+++ src/frontend/apps/impress/src/features/home/components/HomeBottom.tsx
@@ -2,10 +2,10 @@ import { useTranslation } from 'react-i18next';
 
 import IconDocs from '@/assets/icons/icon-docs.svg';
 import { Box, Text } from '@/components';
+import { Title } from '@/components/Title';
 import { useConfig } from '@/core/config';
 import { useCunninghamTheme } from '@/cunningham';
 import { ProConnectButton } from '@/features/auth';
-import { Title } from '@/features/header';
 import { useResponsiveStore } from '@/stores';
 
 export function HomeBottom() {