+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 - ✨(y-provider) preserve callouts, PDFs, page breaks, interlinking
   links and commented text on HTML/markdown export #2296
 - ✨(frontend) add a user menu #2463
+- ✨(frontend) new header and responsive harmonization #2471
 
 ### Changed